add runnable callback examples from old walkthrough tests
