+++ examples/callback-style/get_envelope_info.js
@@ -1,45 +1,20 @@
 // Unit Testing Imports
 var assert = require('assert');
-var fs = require('fs');
 var async = require('async');
 
 var docusign = require('../../docusign.js');
 
-describe('request_signature_on_document', function () {
-  var fullName = 'DocuSign NPM';
+describe('get_envelope_info', function () {
   var docusignEnv = 'demo';
-  var debug = false;
 
-  var config = JSON.parse(fs.readFileSync('config.json'));
-  var integratorKey = config.DOCUSIGN_INTEGRATOR_KEY;
-  var email = config.DOCUSIGN_TEST_EMAIL;
-  var password = config.DOCUSIGN_TEST_PASSWORD;
+  var config = require('../../test-config.json');
+  var debug = config.debug;
+  var integratorKey = config.integratorKey;
+  var email = config.email;
+  var password = config.password;
+  var envelopeId = config.envelopeId;
 
-  var emailSubject = 'DocuSign API - Signature Request on Document Call';
-  var recipients = {};
-
-  recipients.signers = [
-    {
-      'email': email,
-      'name': fullName,
-      'recipientId': 1,
-      'tabs': {
-        'signHereTabs': [{
-          'xPosition': '100',
-          'yPosition': '100',
-          'documentId': '1',
-          'pageNumber': '1'
-        }]
-      }
-    }];
-  var buffer = fs.readFileSync('test/SampleDocument.pdf');
-  var files = [{
-    name: 'SampleDocument.pdf',
-    extension: 'pdf',
-    base64: new Buffer(buffer).toString('base64')
-  }];
-
-  it('should send an envelope with SampleDocument.pdf', function (done) {
+  it('should return envelope information', function (done) {
     async.waterfall([
 
       // **********************************************************************************
@@ -65,12 +40,16 @@ describe('request_signature_on_document', function () {
       },
 
       // **********************************************************************************
-      // Step 3 - Send Envelope
+      // Step 3 - Get Envelope Information
       // **********************************************************************************
-      function sendEnvelope (client, next) {
-        client.envelopes.sendEnvelope(recipients, emailSubject, files, function (err, response) {
-          assert.ok(!err);
-          console.log('The Envelope Information is: \n' + JSON.stringify(response));
+      function getEnvelopeInfo (client, next) {
+        client.envelopes.getEnvelopeInfo(envelopeId, function (error, response) {
+          assert.ok(!error, 'Unexpected ' + error);
+          assert.ok(response.status);
+          assert.ok(response.documentsUri);
+          assert.ok(response.envelopeUri);
+          assert.ok(response.recipientsUri);
+          console.log('The Envelope Information is: \n', response);
           next(null, client);
         });
       },
@@ -82,8 +61,10 @@ describe('request_signature_on_document', function () {
         client.logOut(function (err, response) {
           assert.strictEqual(err, null);
           next(null);
+        // done();
         });
       }
+
     ], function () {
       done();
     });
