+++ package.json
@@ -7,6 +7,7 @@
     "docs": "jsdoc -d doc -r docusign.js dsUtils.js components/ README.md",
     "lint": "semistandard",
     "check-config": "./scripts/check-config.js",
+    "test-cb-examples": "mocha examples/callback-style",
     "test-spec": "ava test/**/*.spec.js",
     "test-cov": "istanbul cover ava -- test/**/*.spec.js",
     "test": "npm run lint && npm run test-cov"
